+++ src/Components/SideNav.tsx
@@ -76,7 +76,8 @@ const reportsNavLinks: INavLink[] = [
 ]
 
   const maintenanceNavLinks: INavLink[] = [
-    { icon: <StorefrontRoundedIcon sx={{ fontSize: '30px'}} />, label: 'Merchants', href: '/system-admin/customer-code' },
+    { icon: <StorefrontRoundedIcon sx={{ fontSize: '30px'}} />, label: 'Merchant', href: '/system-admin/customer-code' },
+    { icon: <StorefrontRoundedIcon sx={{ fontSize: '30px'}} />, label: 'Club', href: '/system-admin/club' },
     { icon: <GroupRoundedIcon sx={{ fontSize: '30px'}} />, label: 'User', href: '/system-admin/user' },
     { icon: <FormatListBulletedRoundedIcon sx={{ fontSize: '30px'}} />, label: 'Analytics', href: '/system-admin/analytics' },
   ]
