+++ src/Pages/Auth/Login.tsx
@@ -120,101 +120,37 @@ const handleSubmitClick = () => {
         setSuccessMessage('Password change successfully');
         setOpenSubmit(false);
         setSubmittedPassword(true);
-        setTimeout(() => {
-          setIsSnackbarOpen(false); 
-            result.RoleId === 1 ? 
-            navigate('accounting/dashboard-accounting') :  
-            result.RoleId === 2 ? 
-            navigate('treasury/dashboard-treasury') : 
-            result.RoleId === 4 ? 
-            navigate('system-admin/dashboard-system-admin') : 
-            navigate('maintenance')
-          window.location.reload()
-        }, 1000,);
-      }
-      else
-      {
-        setIsSnackbarOpen(true);
-        setSnackbarSeverity('error');
-        setErrorMessage('Error password change. Please try again!');
-        setOpenSubmit(false);
-        setSubmittedPassword(true);
-      }
-    })
-    .catch((error) => {
-      setIsSnackbarOpen(true);
-      setSnackbarSeverity('error');
-      setErrorMessage('Error password change');
-    })
-  } catch (error) {
-      setIsSnackbarOpen(true);
-      setSnackbarSeverity('error');
-      setErrorMessage('Error password change');
-  } 
-};
-
-const handleClickShowPassword = () => setShowPassword((show) => !show);
-
-// Handle closing the snackbar
-const handleCloseSubmit = () => {
-  setOpenSubmit(false);
-};
-
-const handleSubmitClick = () => {
-  setSubmittedPassword(true);
-  try {
-    const updatedParam: IFirstLogin = {
-      Username: login.Username,
-      Password: confirmPassword
-    }
-
-    const submitChangePassword: AxiosRequestConfig = {
-      method: 'POST',
-      url: `${REACT_APP_API_ENDPOINT}/Auth/ChangePassword`,
-      data: updatedParam,
-    };
-
-    axios(submitChangePassword)
-    .then(async (response) => {
-      var result = response.data;
-      if(result.Message === 'Successful') 
-      {
-        setIsSnackbarOpen(true);
-        setSnackbarSeverity('success');
-        setSuccessMessage('Password change successfully');
-        setOpenSubmit(false);
-        setSubmittedPassword(true);
-        setTimeout(() => {
-          setIsSnackbarOpen(false); 
-            result.RoleId === 1 ? 
-            navigate('accounting/dashboard-accounting') :  
-            result.RoleId === 2 ? 
-            navigate('treasury/dashboard-treasury') : 
-            result.RoleId === 4 ? 
-            navigate('system-admin/dashboard-system-admin') : 
-            navigate('maintenance')
-          window.location.reload()
-        }, 1000,);
-      }
-      else
-      {
-        setIsSnackbarOpen(true);
-        setSnackbarSeverity('error');
-        setErrorMessage('Error password change. Please try again!');
-        setOpenSubmit(false);
-        setSubmittedPassword(true);
-      }
-    })
-    .catch((error) => {
-      setIsSnackbarOpen(true);
-      setSnackbarSeverity('error');
-      setErrorMessage('Error password change');
-    })
-  } catch (error) {
-      setIsSnackbarOpen(true);
-      setSnackbarSeverity('error');
-      setErrorMessage('Error password change');
-  } 
+setTimeout(() => {
+  setIsSnackbarOpen(false); 
+    result.RoleId === 1 ? 
+    navigate('accounting/dashboard-accounting') :  
+    result.RoleId === 2 ? 
+    navigate('treasury/dashboard-treasury') : 
+    result.RoleId === 4 ? 
+    navigate('system-admin/dashboard-system-admin') : 
+    navigate('maintenance')
+  window.location.reload()
+}, 1000,);
+}
+else
+{
+setIsSnackbarOpen(true);
+setSnackbarSeverity('error');
+setErrorMessage('Error password change. Please try again!');
+setOpenSubmit(false);
+setSubmittedPassword(true);
+}
+})
+.catch((error) => {
+setIsSnackbarOpen(true);
+setSnackbarSeverity('error');
+setErrorMessage('Error password change');
+})
+} catch (error) {
+setIsSnackbarOpen(true);
+setSnackbarSeverity('error');
+setErrorMessage('Error password change');
+} 
 };
 
 const handleClickShowPassword = () => setShowPassword((show) => !show);
@@ -262,30 +198,6 @@ const handleLoginSubmit = () => {
           }, 1000,);
         }
       }
-        auth.signIn(result);
-        if(result.IsFirstLogin === true)
-        {
-          setOpenSubmit(true)
-        }
-        else
-        {
-          setIsSnackbarOpen(true);
-          setSnackbarSeverity('success');
-          setSuccessMessage('Login successfully!')
-          setSubmitted(true);
-          setTimeout(() => {
-            setIsSnackbarOpen(false); 
-              result.RoleId === 1 ? 
-              navigate('accounting/dashboard-accounting') :  
-              result.RoleId === 2 ? 
-              navigate('treasury/dashboard-treasury') : 
-              result.RoleId === 4 ? 
-              navigate('system-admin/dashboard-system-admin') : 
-              navigate('maintenance')
-            window.location.reload()
-          }, 1000,);
-        }
-      }
       else if(result.Message === 'Login attempt limit reached!')
         {
           setIsSnackbarOpen(true);
@@ -296,6 +208,7 @@ const handleLoginSubmit = () => {
             Password: ""
           })
         }
+      
       else if(result.Message === 'User not found.')
       {
         setIsSnackbarOpen(true);
