+++ src/Pages/CustomerCode/CustomerCode.tsx
@@ -1,4 +1,4 @@
-import { Box, Grid, IconButton, Table, TableBody, TableCell, TableHead, TableRow, Typography, styled, TextField, InputAdornment, MenuItem, CircularProgress, Snackbar, Fade, Alert, Pagination, DialogContentText, Paper, Divider } from '@mui/material';
+import { Box, Grid, IconButton, Table, TableBody, TableCell, TableHead, TableRow, Typography, styled, TextField, InputAdornment, MenuItem, CircularProgress, Snackbar, Fade, Alert, Pagination, DialogContentText, Paper, Divider, Button } from '@mui/material';
 import { useCallback, useEffect, useState } from 'react';
 import { Search as SearchIcon,} from '@mui/icons-material/';
 import axios, { AxiosRequestConfig } from 'axios';
@@ -11,24 +11,30 @@ import DeleteIcon from '@mui/icons-material/Delete';
 import EditIcon from '@mui/icons-material/Edit';
 import IAnalyticProps from '../Common/Interface/IAnalyticsProps';
 import { insertLogs } from '../../Components/Functions/InsertLogs';
+import AddIcon from '@mui/icons-material/Add';
 
 const StyledTableCellHeader = styled(TableCell)(() => ({
-  fontSize: "15px",
-  fontWeight: '900',
+  fontSize: '20px',
+  fontWeight: '100',
+  padding: '10px',
   color: '#1C2C5A',
-  textAlign: 'center'
 }));
 
 const StyledTableCellBody = styled(TableCell)(() => ({
-  fontSize: "12px",
-  padding: "1px",
+  fontSize: '11px',
+  fontWeight: '900',
+  padding: '1px',
   color: '#1C2C5A',
-  textAlign: 'center'
 })); 
+
+const StyledTextField = styled(TextField)(() => ({
+  fontSize: '10px',
+  fontWeight: '100',
+}))
 
 const BootstrapButton = styled(IconButton)(() => ({
   textTransform: 'none',
-  fontSize: 12, 
+  fontSize: 11, 
   lineHeight: 1.5,
   color: '#1C2C5A',
   fontWeight: '900',
@@ -64,16 +70,26 @@ const WhiteAlert = styled(Alert)(({ severity }) => ({
   backgroundColor: severity === 'success' ? '#E7FFDF' : '#FFC0C0',
 }));
 
+  // Create a mapping object
+  const defaultFormValue = {
+    CustomerName: "",
+    CustomerCode: "",
+    DeleteFlag: false,
+    CustomerNo: "",
+    UserId: "",
+  };
+
 const CustomerCode = () => {
   const { REACT_APP_API_ENDPOINT } = process.env;
   const [customerCodes, setCustomerCodes] = useState<ICustomerCode[]>([]);
-  const [fieldValues, setFieldValues] = useState<ICustomerCodeUpdateDelete>();
+  const [fieldValues, setFieldValues] = useState<ICustomerCodeUpdateDelete>(defaultFormValue);
   const [loading, setLoading] = useState<boolean>(true);
   const [snackbarSeverity, setSnackbarSeverity] = useState<'error' | 'warning' | 'info' | 'success'>('success'); // Snackbar severity
   const [isSnackbarOpen, setIsSnackbarOpen] = useState<boolean>(false); // Snackbar open state
   const [message, setMessage] = useState<string>(''); // Error message
   const [openEdit, setOpenEdit] = useState<boolean>(false);
-  const [openDelete, setOpenDelete] = useState<boolean>(false);
+  const [openInsert, setOpenInsert] = useState<boolean>(false);
+  const [isDisabled, setIsDisabled] = useState<boolean>(false);
   const [submitted, setSubmitted] = useState<boolean>(false); // Form submission status
   const [searchQuery, setSearchQuery] = useState<string>(''); // Search query
   const [page, setPage] = useState<number>(1); // Current page number
@@ -111,12 +127,52 @@ const CustomerCode = () => {
   const handleCloseModalEdit = useCallback(() => {
     setOpenEdit(false);
     setSubmitted(false);
+    setFieldValues(defaultFormValue);
   }, []);
 
-  const handleCloseModalDelete = useCallback(() => {
-    setOpenDelete(false);
+  const handleCloseModalInsert = useCallback(() => {
+    setOpenInsert(false);
     setSubmitted(false);
+    setFieldValues(defaultFormValue);
   }, []);
+
+  const handleSubmitInsert = async() => {
+    setSubmitted(true);
+    const allIsPopulated = !fieldValues.CustomerCode || !fieldValues.CustomerName || !fieldValues.CustomerNo;
+    
+    if (allIsPopulated) {
+      setIsSnackbarOpen(true);
+      setSnackbarSeverity('error');
+      setMessage('Please input required field.');
+      return;
+    }
+    
+    const insertRequest: AxiosRequestConfig = {
+      method: 'POST',
+      url: `${REACT_APP_API_ENDPOINT}/CustomerCode/InsertCustomerCodeAsync`,
+      data: fieldValues,
+    };
+
+    axios(insertRequest)
+      .then(() => {
+        setSubmitted(false);
+        setIsSnackbarOpen(true);
+        setSnackbarSeverity('success');
+        setMessage('Customer insert successfully!')
+        fetchCustomerCodes(page, itemsPerPage, searchQuery, columnToSort, orderBy);
+        handleCloseModalInsert();
+        setSearchQuery('');
+      })
+      .catch((error) => {
+        setIsSnackbarOpen(true);
+        setSnackbarSeverity('error');
+        if (error.response && error.response.status === 404) {
+          setMessage('Customer already exist on database');
+        } else {
+          setMessage('Error occurred. Please try again.');
+        }
+      });
+  }
 
   const fetchCustomerCodes = useCallback(async(pageNumber: number, pageSize: number, searchQuery: string | null, columnToSort: string | null, orderBy: string | null) => {
     try {
@@ -156,24 +212,6 @@ const CustomerCode = () => {
   }, [fetchCustomerCodes, page, itemsPerPage, searchQuery, columnToSort, orderBy]);
 
   // Handle changes in form fields
-  const handleChangeCustomer = (field: keyof ICustomerCodeUpdateDelete, value: any) => {
-    setFieldValues((prevValues) => ({
-      ...prevValues,
-      [field]: value,
-      UserId: Id
-    }) as ICustomerCodeUpdateDelete | undefined);
-  };
-
-  const handleEditOrDeleteClick = (id: number, isEdit: boolean) => {
-    const customerCodeToUpdate = customerCodes.find(code => code.Id === id);
-    if (customerCodeToUpdate) {
-      setFieldValues(customerCodeToUpdate);
-      isEdit ? setOpenEdit(true) : setOpenDelete(true);
-    } else {
-      console.error(`Customer code with id ${id} not found.`);
-    }
-  };
-
   const handleSubmitEdit = async() => {
     setSubmitted(true);
     const allIsPopulated = !fieldValues?.CustomerNo || !fieldValues.CustomerName || !fieldValues.CustomerCode;
@@ -199,6 +237,7 @@ const CustomerCode = () => {
         setMessage('Customer successfully updated!')
         fetchCustomerCodes(page, itemsPerPage, searchQuery, columnToSort, orderBy);
         handleCloseModalEdit();
+        setSearchQuery('');
       })
       .catch((error) => {
         console.error("Error updating data:", error);
@@ -209,35 +248,27 @@ const CustomerCode = () => {
   }
 
   // Handle form submission for deleting a zone type
-  const handleSubmitDelete = async () => {
-    const deleteRequest: AxiosRequestConfig = {
-      method: 'PUT',
-      url: `${REACT_APP_API_ENDPOINT}/CustomerCode/DeleteCustomerCodeByIdAsync?id=${fieldValues?.Id}`,
-    };
+  // -------------------------------
+  //Handle the Display of Create/Update Modals
+  const handleEditOrInsertClick = (ID?: number) => {
+    const customerToUpdate = customerCodes.find(customer => customer.Id === ID);
+    if (customerToUpdate) 
+    {
+        setFieldValues(customerToUpdate);
+        setOpenEdit(true);
+    }
+    else 
+    {
+        setOpenInsert(true);
+    }
+  };
 
-    await axios(deleteRequest)
-      .then(async () => {
-        setIsSnackbarOpen(true);
-        setSnackbarSeverity('success');
-        setMessage('Data deleted successfully!')
-        handleCloseModalDelete();
-  
-        const anaylticsParamUpdated: IAnalyticProps = {
-          userId: Id,
-          action: 'Delete Merchant',
-          remarks: `Id: ${fieldValues?.Id} : Successfully Deleted`
-        };
-
-        await insertLogs(anaylticsParamUpdated);
-
-        fetchCustomerCodes(1, itemsPerPage, "", "", "");
-      })
-      .catch((error) => {
-        console.error("Error saving data:", error);
-        setIsSnackbarOpen(true);
-        setSnackbarSeverity('error');
-        setMessage('Error occurred. Please try again.');
-      });
+  // Handle changes in form fields
+  const handleChangeCustomerUpdate = (field: keyof ICustomerCodeUpdateDelete, value: any) => {
+    setFieldValues((prevValues) => ({
+      ...prevValues,
+      [field]: value
+    }) as ICustomerCodeUpdateDelete);
   };
 
   if (!loading) {
@@ -250,127 +281,119 @@ const CustomerCode = () => {
         flexGrow: 1,
       }}
     >
-      <Paper elevation={3} sx={{ padding: '20px', maxWidth: '100%', borderRadius: '15px', height: '750px' }}>
-        <Typography variant="h5" gutterBottom sx={{ fontWeight: 'bold', marginBottom: '10px', color: '#1C2C5A', }}>
-          Merchants
-        </Typography>
-        <Divider sx={{ marginBottom: '20px' }} />
-        <Grid container spacing={1} sx={{ height: '60px' }}>
-          <Grid item xs={12} sm={4}>
-            <TextField
-              placeholder='Search'
-              size='small'
-              fullWidth
-              InputProps={{
-                sx: {
-                  borderRadius: '20px', 
-                  backgroundColor: '#EEEEEE',
-                  color: '#1C2C5A',
-                  "& fieldset": { border: 'none' },
-                  boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
-                },
-                endAdornment: (
-                  <InputAdornment position="end">
-                    <SearchIcon />
-                  </InputAdornment>
-                ),
-              }}
-              variant="outlined"
-              value={searchQuery}
-              onChange={handleSearchChange}
-              autoFocus 
-            />
-          </Grid>
+      <Typography variant="h5" gutterBottom sx={{ fontWeight: 'bold', marginBottom: '10px', color: '#1C2C5A', }}>
+        Merchants
+      </Typography>
+      <Grid container spacing={1} sx={{ height: '60px' }}>
+        <Grid item xs={12} sm={4}>
+          <StyledTextField
+            placeholder='Search'
+            size='small'
+            fullWidth
+            InputProps={{
+              sx: {
+                fontSize: '11px',
+                borderRadius: '20px', 
+                backgroundColor: '#EEEEEE',
+                color: '#1C2C5A',
+                "& fieldset": { border: 'none' },
+                boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
+              },
+              endAdornment: (
+                <InputAdornment position="end">
+                  <SearchIcon />
+                </InputAdornment>
+              ),
+            }}
+            variant="outlined"
+            value={searchQuery}
+            onChange={handleSearchChange}
+            autoFocus 
+          />
         </Grid>
-        <Divider />
-          <CustomScrollbarBox component={Paper}
-            sx={{
-              height: '600px',
-              position: 'relative',
-              paddingTop: '10px',
-              borderBottomLeftRadius: '20px',
-              borderBottomRightRadius: '20px',
-              borderTopLeftRadius: '0',
-              borderTopRightRadius: '0',
-              boxShadow: 'none',
-              paddingLeft: '20px',
-              paddingRight: '20px',
-            }}
-          >
-          <Table
-            sx={{
-              backgroundColor: '#ffffff',
-            }}
-            aria-label="spanning table">
-            <TableHead
+        <Grid item xs={12} sm={6.5}/>
+        <Grid item xs={12} sm={1.2}>
+            <Button
+              fullWidth
+              onClick={() => {
+                handleEditOrInsertClick(undefined)
+              }}
               sx={{
-                zIndex: 3,
-                position: 'sticky',
-                top: '-10px',
-                backgroundColor: '#ffffff',
+                textTransform: 'none',
+                backgroundColor: "#4761AD",
+                '&:hover': {
+                  backgroundColor: "#20346E",
+                  color: "#FFFFFF",
+                },
+                color: "#FFFFFF",
+                fontWeight: 'bold',
+                fontSize: '11px',
+                height: '40px',
+                borderRadius: '15px',
+                boxShadow: '1px 5px 4px -1px rgba(0,0,0,0.3)',
               }}
             >
-                <TableRow>
-                  <StyledTableCellHeader sx={{ width: '185px'}}>ID</StyledTableCellHeader>
-                  <StyledTableCellHeader>Customer Name</StyledTableCellHeader>
-                  <StyledTableCellHeader sx={{ width: '200px'}}>Code</StyledTableCellHeader>
-                  <StyledTableCellHeader>Customer No</StyledTableCellHeader>
-                  <StyledTableCellHeader>Action</StyledTableCellHeader>
-                </TableRow>
-              </TableHead>
-              <TableBody >
-                {customerCodes.map((row, index) => (
-                <TableRow key={index} sx={{ "& td": { border: 0 }}}>
-                  <StyledTableCellBody sx={{ textAlign: 'center', width: '160px' }} >{row.Id}</StyledTableCellBody>
-                  <StyledTableCellBody sx={{ textAlign: 'start', width: '320px' }}>{row.CustomerName}</StyledTableCellBody>
-                  <StyledTableCellBody sx={{ textAlign: 'center', width: '170px' }}>{row.CustomerCode}</StyledTableCellBody>
-                  <StyledTableCellBody sx={{ textAlign: 'center', width: '250px' }}>{row.CustomerNo}</StyledTableCellBody>
-                  <StyledTableCellBody sx={{ textAlign: 'center', width: '60px' }}>
-                    <Box sx={{ display: 'flex', justifyContent: 'center' }}>
-                      <BootstrapButton
+              <AddIcon/>Add Merchant
+            </Button>
+          </Grid>
+      </Grid>
+      <Paper elevation={3} sx={{ backgroundcolor: '#B0B0B0', padding: '10px', maxWidth: '100%', borderRadius: '15px', height: '690px', display: 'flex', flexDirection: 'column' }}>
+        <Table
+          sx={{
+            backgroundColor: '#ffffff',
+          }}
+          aria-label="spanning table">
+          <TableHead
+            sx={{
+              zIndex: 3,
+              position: 'sticky',
+              top: '-10px',
+              backgroundColor: '#ffffff',
+            }}
+          >
+            <TableRow sx={{ borderBottom: "solid  #20346E" }}>
+              <StyledTableCellHeader sx={{ textAlign: 'left' }} onClick={() => setColumnToSort('CustomerCode')}>Code</StyledTableCellHeader>
+              <StyledTableCellHeader sx={{ textAlign: 'left' }} onClick={() => setColumnToSort('CustomerName')}>Merchant</StyledTableCellHeader>
+              <StyledTableCellHeader sx={{ textAlign: 'left' }} onClick={() => setColumnToSort('CustomerNo')}>Customer No</StyledTableCellHeader>
+              <StyledTableCellHeader sx={{ textAlign: 'left' }} onClick={() => setColumnToSort('DeleteFlag')}>Status</StyledTableCellHeader>
+              <StyledTableCellHeader sx={{ textAlign: 'center' }}>Action</StyledTableCellHeader>
+            </TableRow>
+          </TableHead>
+          <TableBody >
+            {customerCodes.map((row, index) => (
+              <TableRow key={index} sx={{ "& td": { border: 0 }}}>
+                <StyledTableCellBody sx={{ textAlign: 'left' }}>{row.CustomerCode}</StyledTableCellBody>
+                <StyledTableCellBody sx={{ textAlign: 'left' }}>{row.CustomerName}</StyledTableCellBody>
+                <StyledTableCellBody sx={{ textAlign: 'left' }}>{row.CustomerNo}</StyledTableCellBody>
+                <StyledTableCellBody sx={{ textAlign: 'left' }}>{row.DeleteFlag ? 'Inactive' : 'Active'}</StyledTableCellBody>
+                <StyledTableCellBody sx={{ textAlign: 'left' }}>
+                  <Box sx={{ display: 'flex', justifyContent: 'center' }}>
+                  <BootstrapButton
                         onClick={() => {
-                          handleEditOrDeleteClick(row.Id, true)
+                          handleEditOrInsertClick(row.Id)
                         }}
                         sx={{
-                          backgroundColor: "#FCBA70",
+                          backgroundColor: "#4761AD",
                           width: '90px',
                           height: "20px",
                           borderRadius: "15px",
-                          color:"#634422",
+                          color: "#FFFFFF",
                           marginLeft: 0.5,
                           "&:hover": {
-                            backgroundColor: "#FF9419",
-                            color: '#FFFFFF',
+                            backgroundColor: "#20346E",
+                            color: "#FFFFFF",
                           },
-                        }}>
-                        <EditIcon  sx={{fontSize: '15px', marginRight: '2px'}}/>Edit
+                        }}>View
                       </BootstrapButton>
-                      <BootstrapButton
-                        onClick={() => {
-                          handleEditOrDeleteClick(row.Id, false)
-                        }}
-                        sx={{
-                          backgroundColor: "#FFB5B5",
-                          width: '90px',
-                          height: "20px",
-                          borderRadius: "15px",
-                          color: "#644848",
-                          marginLeft: 0.5,
-                          "&:hover": {
-                            backgroundColor: "#FF7171",
-                            color: '#FFFFFF',
-                          },
-                        }}>
-                        <DeleteIcon  sx={{fontSize: '15px', marginRight: '2px'}}/>Delete
-                      </BootstrapButton>
-                    </Box>
-                  </StyledTableCellBody>
-                </TableRow>
-  
-                ))}
-              </TableBody>
-            </Table>
-          </CustomScrollbarBox>
+                  </Box>
+                </StyledTableCellBody>
+              </TableRow>
+            ))}
+            </TableBody>
+          </Table>
+          <Box sx={{ flex: '1 0 auto' }}>
+            {/* Content that grows to fill the space */}
+          </Box>
           {/* Pagination */}
           <Box sx={{ display: 'flex', justifyContent: 'flex-end', mt: 0 }}>
             <Pagination
@@ -385,7 +408,7 @@ const CustomerCode = () => {
             />
           </Box>
       </Paper>
-        {/* Snackbar for displaying messages */}
+      {/* Snackbar for displaying messages */}
         <Snackbar
           open={isSnackbarOpen}
           autoHideDuration={3000}
@@ -400,38 +423,75 @@ const CustomerCode = () => {
             {message}
           </WhiteAlert>
         </Snackbar>
-      <ModalComponent
-        title='Edit Customer'
+        <ModalComponent
+        title='Merchant Details'
         onClose={handleCloseModalEdit}
-        buttonName='Save'
+        buttonName='Save Changes'
         open={openEdit}
         onSave={handleSubmitEdit}
         children={
           <Box sx={{ flexGrow: 1 }}>
             <Grid container spacing={1}>
-              <Grid item xs={8}
+              <Grid item xs={12}
                 sx={{
                   fontFamily: 'Inter',
                   fontWeight: '900',
                   color: '#1C2C5A',
-                  fontSize: '20px'
+                  fontSize: '12px'
                 }}>
-                Customer Name *
+                Merchant Code *
               </Grid>
-              <Grid item xs={11.5} sx={{marginLeft: '10px'}}>
+              <Grid item xs={12}>
                 <Box display={'flex'}>
-                  <TextField 
+                  <StyledTextField
+                    fullWidth
+                    variant="outlined"
+                    size="small"
+                    name="FirstName"
+                    type="text"
+                    required
+                    value={fieldValues?.CustomerCode}
+                    onChange={(e) => handleChangeCustomerUpdate("CustomerCode", e.target.value.trim() === ''? '' : e.target.value)}
+                    error={submitted && !fieldValues?.CustomerCode}
+                    helperText={submitted && !fieldValues?.CustomerCode && "Merchant Code is required"}
+                    InputProps={{
+                      sx: {
+                        fontSize: '11px', 
+                        borderRadius: '13px', 
+                        backgroundColor: '#EEEEEE',
+                        color: '#1C2C5A',
+                        "& fieldset": { border: 'none' },
+                        boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
+                      },
+                    }}
+                  >
+                  </StyledTextField>
+                </Box>
+              </Grid>
+              <Grid item xs={12}
+                sx={{
+                  fontFamily: 'Inter',
+                  fontWeight: '900',
+                  color: '#1C2C5A',
+                  fontSize: '12px'
+                }}>
+                Merchant Name *
+              </Grid>
+              <Grid item xs={12}>
+                <Box display={'flex'}>
+                  <StyledTextField 
                     size='small' 
                     type="text"
                     fullWidth
                     variant="outlined"
                     required
                     value={fieldValues?.CustomerName}
-                    onChange={(e) => handleChangeCustomer("CustomerName", e.target.value.trim() === ''? '' : e.target.value)}
+                    onChange={(e) => handleChangeCustomerUpdate("CustomerName", e.target.value.trim() === ''? '' : e.target.value)}
                     error={submitted && !fieldValues?.CustomerName}
-                    helperText={submitted && !fieldValues?.CustomerName && "Customer Name is required"}
+                    helperText={submitted && !fieldValues?.CustomerName && "Merchant Name is required"}
                     InputProps={{
                       sx: {
+                        fontSize: '11px', 
                         borderRadius: '13px', 
                         backgroundColor: '#EEEEEE',
                         color: '#1C2C5A',
@@ -440,67 +500,33 @@ const CustomerCode = () => {
                       },
                     }}
                   >
-                  </TextField>
+                  </StyledTextField>
                 </Box>
               </Grid>
-              <Grid item xs={8}
+              <Grid item xs={12}
                 sx={{
                   fontFamily: 'Inter',
                   fontWeight: '900',
                   color: '#1C2C5A',
-                  fontSize: '20px'
+                  fontSize: '12px'
                 }}>
-                Customer Code *
+                Merchant No. *
               </Grid>
-              <Grid item xs={11.5} sx={{marginLeft: '10px'}}>
+              <Grid item xs={12}>
                 <Box display={'flex'}>
-                  <TextField 
-                    type="text"
+                  <StyledTextField 
                     size='small' 
-                    fullWidth
-                    variant="outlined"
-                    required
-                    value={fieldValues?.CustomerCode}
-                    onChange={(e) => handleChangeCustomer("CustomerCode", e.target.value.trim() === ''? '' : e.target.value)}
-                    error={submitted && !fieldValues?.CustomerCode}
-                    helperText={submitted && !fieldValues?.CustomerCode && "Customer Code is required"}
-                    InputProps={{
-                      sx: {
-                        borderRadius: '13px', 
-                        backgroundColor: '#EEEEEE',
-                        color: '#1C2C5A',
-                        "& fieldset": { border: 'none' },
-                        boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
-                      },
-                    }}
-                  >
-                  </TextField>
-                </Box>
-              </Grid>
-              <Grid item xs={8}
-                sx={{
-                  fontFamily: 'Inter',
-                  fontWeight: '900',
-                  color: '#1C2C5A',
-                  fontSize: '20px'
-                }}>
-                Customer No *
-              </Grid>
-              <Grid item xs={11.5} sx={{marginLeft: '10px'}}>
-                <Box display={'flex'}>
-                  <TextField
-                    fullWidth
-                    variant="outlined"
-                    size="small"
-                    name="CustomerNo"
                     type="text"
+                    fullWidth
+                    variant="outlined"
                     required
                     value={fieldValues?.CustomerNo}
-                    onChange={(e) => handleChangeCustomer("CustomerNo", e.target.value.trim() === ''? '' : e.target.value)}
+                    onChange={(e) => handleChangeCustomerUpdate("CustomerNo", e.target.value.trim() === ''? '' : e.target.value)}
                     error={submitted && !fieldValues?.CustomerNo}
-                    helperText={submitted && !fieldValues?.CustomerNo && "Category is required"}
+                    helperText={submitted && !fieldValues?.CustomerNo && "Merchant No. is required"}
                     InputProps={{
                       sx: {
+                        fontSize: '11px', 
                         borderRadius: '13px', 
                         backgroundColor: '#EEEEEE',
                         color: '#1C2C5A',
@@ -509,7 +535,44 @@ const CustomerCode = () => {
                       },
                     }}
                   >
-                  </TextField>
+                  </StyledTextField>
+                </Box>
+              </Grid>
+              <Grid item xs={12}
+                sx={{
+                  fontFamily: 'Inter',
+                  fontWeight: '900',
+                  color: '#1C2C5A',
+                  fontSize: '12px'
+                }}>
+                Status
+              </Grid>
+              <Grid item xs={12}>
+                <Box display={'flex'}>
+                  <StyledTextField 
+                    size='small' 
+                    type="text"
+                    fullWidth
+                    variant="outlined"
+                    required
+                    select
+                    value={fieldValues?.DeleteFlag}
+                    onChange={(e) => handleChangeCustomerUpdate("DeleteFlag", e.target.value.trim() === ''? '' : e.target.value)}
+                    error={submitted && !fieldValues?.DeleteFlag}
+                    InputProps={{
+                      sx: {
+                        fontSize: '11px',
+                        borderRadius: '13px', 
+                        backgroundColor: '#EEEEEE',
+                        color: '#1C2C5A',
+                        "& fieldset": { border: 'none' },
+                        boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
+                      },
+                    }}
+                  >
+                    <MenuItem sx={{ color: '#1C2C5A', fontSize: '11px', }} key={'false'} value={'false'}>Active</MenuItem>
+                    <MenuItem sx={{ color: '#1C2C5A', fontSize: '11px', }} key={'true'} value={'true'}>Inactive</MenuItem>
+                  </StyledTextField>
                 </Box>
               </Grid>
             </Grid>
@@ -517,20 +580,158 @@ const CustomerCode = () => {
         } 
       />
       <ModalComponent
-        title='Warning!'
-        onClose={handleCloseModalDelete}
-        buttonName='Yes'
-        open={openDelete}
-        onSave={handleSubmitDelete}
+        title='Add Merchant'
+        onClose={handleCloseModalInsert}
+        buttonName='Save'
+        open={openInsert}
+        onSave={handleSubmitInsert}
         children={
           <Box sx={{ flexGrow: 1 }}>
-            <DialogContentText
-              sx={{ color: "#1C2C5A", textAlign: 'center', fontSize: '20px', fontWeight: '900' }}
-            >
-              This action cannot be undone,
-              <br />
-              Continue Delete?
-            </DialogContentText>
+            <Grid container spacing={1}>
+              <Grid item xs={12}
+                sx={{
+                  fontFamily: 'Inter',
+                  fontWeight: '900',
+                  color: '#1C2C5A',
+                  fontSize: '12px'
+                }}>
+                Merchant Code *
+              </Grid>
+              <Grid item xs={12}>
+                <Box display={'flex'}>
+                  <StyledTextField
+                    fullWidth
+                    variant="outlined"
+                    size="small"
+                    name="FirstName"
+                    type="text"
+                    required
+                    value={fieldValues?.CustomerCode}
+                    onChange={(e) => handleChangeCustomerUpdate("CustomerCode", e.target.value.trim() === ''? '' : e.target.value)}
+                    error={submitted && !fieldValues?.CustomerCode}
+                    helperText={submitted && !fieldValues?.CustomerCode && "Merchant Code is required"}
+                    InputProps={{
+                      sx: {
+                        fontSize: '11px', 
+                        borderRadius: '13px', 
+                        backgroundColor: '#EEEEEE',
+                        color: '#1C2C5A',
+                        "& fieldset": { border: 'none' },
+                        boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
+                      },
+                    }}
+                  >
+                  </StyledTextField>
+                </Box>
+              </Grid>
+              <Grid item xs={12}
+                sx={{
+                  fontFamily: 'Inter',
+                  fontWeight: '900',
+                  color: '#1C2C5A',
+                  fontSize: '12px'
+                }}>
+                Merchant Name *
+              </Grid>
+              <Grid item xs={12}>
+                <Box display={'flex'}>
+                  <StyledTextField 
+                    size='small' 
+                    type="text"
+                    fullWidth
+                    variant="outlined"
+                    required
+                    value={fieldValues?.CustomerName}
+                    onChange={(e) => handleChangeCustomerUpdate("CustomerName", e.target.value.trim() === ''? '' : e.target.value)}
+                    error={submitted && !fieldValues?.CustomerName}
+                    helperText={submitted && !fieldValues?.CustomerName && "Merchant Name is required"}
+                    InputProps={{
+                      sx: {
+                        fontSize: '11px', 
+                        borderRadius: '13px', 
+                        backgroundColor: '#EEEEEE',
+                        color: '#1C2C5A',
+                        "& fieldset": { border: 'none' },
+                        boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
+                      },
+                    }}
+                  >
+                  </StyledTextField>
+                </Box>
+              </Grid>
+              <Grid item xs={12}
+                sx={{
+                  fontFamily: 'Inter',
+                  fontWeight: '900',
+                  color: '#1C2C5A',
+                  fontSize: '12px'
+                }}>
+                Merchant No. *
+              </Grid>
+              <Grid item xs={12}>
+                <Box display={'flex'}>
+                  <StyledTextField 
+                    size='small' 
+                    type="text"
+                    fullWidth
+                    variant="outlined"
+                    required
+                    value={fieldValues?.CustomerNo}
+                    onChange={(e) => handleChangeCustomerUpdate("CustomerNo", e.target.value.trim() === ''? '' : e.target.value)}
+                    error={submitted && !fieldValues?.CustomerNo}
+                    helperText={submitted && !fieldValues?.CustomerNo && "Merchant No. is required"}
+                    InputProps={{
+                      sx: {
+                        fontSize: '11px', 
+                        borderRadius: '13px', 
+                        backgroundColor: '#EEEEEE',
+                        color: '#1C2C5A',
+                        "& fieldset": { border: 'none' },
+                        boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
+                      },
+                    }}
+                  >
+                  </StyledTextField>
+                </Box>
+              </Grid>
+              <Grid item xs={12}
+                sx={{
+                  fontFamily: 'Inter',
+                  fontWeight: '900',
+                  color: '#1C2C5A',
+                  fontSize: '12px'
+                }}>
+                Status
+              </Grid>
+              <Grid item xs={12}>
+                <Box display={'flex'}>
+                  <StyledTextField 
+                    size='small' 
+                    type="text"
+                    fullWidth
+                    variant="outlined"
+                    required
+                    select
+                    value={fieldValues?.DeleteFlag}
+                    onChange={(e) => handleChangeCustomerUpdate("DeleteFlag", e.target.value.trim() === ''? '' : e.target.value)}
+                    error={submitted && !fieldValues?.DeleteFlag}
+                    InputProps={{
+                      sx: {
+                        fontSize: '11px',
+                        borderRadius: '13px', 
+                        backgroundColor: '#EEEEEE',
+                        color: '#1C2C5A',
+                        "& fieldset": { border: 'none' },
+                        boxShadow: 'inset 1px 1px 1px -3px rgba(0,0,0,0.1), inset 1px 1px 8px 0px rgba(0,0,0,0.3)',
+                      },
+                    }}
+                  >
+                    <MenuItem sx={{ color: '#1C2C5A', fontSize: '11px', }} key={'false'} value={'false'}>Active</MenuItem>
+                    <MenuItem sx={{ color: '#1C2C5A', fontSize: '11px', }} key={'true'} value={'true'}>Inactive</MenuItem>
+                  </StyledTextField>
+                </Box>
+              </Grid>
+            </Grid>
           </Box>
         } 
       />
