+++ src/Pages/CustomerCode/Interface/ICustomerUpdateDelete.ts
@@ -1,7 +1,7 @@
 export default interface ICustomerCodeUpdateDelete {
-  Id: number;
-  CustomerName: string;
-  CustomerCode: string;
+  Id?: number;
+  CustomerName?: string;
+  CustomerCode?: string;
   DeleteFlag?: boolean;
   CustomerNo?: string;
   UserId?: string;
