+++ src/Pages/DashboardStores/Interface/IDashboardStores.ts
@@ -1,0 +1,8 @@
+export default interface IDashboardStores {
+    Id: number
+    LocationCode: number
+    LocationName: string
+    ShortName: string
+    DeleteFlag: boolean
+  }
+
