+++ src/Pages/DashboardStores/Interface/IStoreUpdateDelete.ts
@@ -1,0 +1,8 @@
+export default interface IStoreUpdateDelete {
+    Id?: number
+    LocationCode?: number
+    LocationName?: string
+    ShortName?: string
+    DeleteFlag?: boolean
+  }
+
